requirements.txt
python-i18n==0.3.7